bug-erms-1583
as of ERMS-1583, the adding/removing from issue entitlement coverage statements has been moved to SubscriptionController
