+++ app/application.properties
@@ -1,7 +1,7 @@
 #Grails Metadata file
-#Thu Aug 22 11:05:01 CEST 2019
-app.buildDate=22.08.2019 11\:05
-app.buildNumber=7219
+#Thu Aug 22 13:25:40 CEST 2019
+app.buildDate=22.08.2019 13\:25
+app.buildNumber=7221
 app.buildProfile=development
 app.grails.version=2.5.6
 app.name=laser
